+++ Changelog.md
@@ -1,1 +1,2 @@
 # Changelog
+- Don't destroy NEO when a transport pod is travelling towards it or a caravan is on there